Updated testing of all possible results. Now the game resets after the execution of a command, before attempting to execute the next command.
Moved game_elements_repository dictionary to the Game class. Now the Game class also implements the singleton pattern, so only one game instance exists at a time and is accessible from other modules when needed.

Implemented a reset method in Game class to re-initialize the instance to the defaults values.